Changes related to renaming some concept.

+++ src/main/java/com/inncretech/multitenancy/datasource/config/RoutingDataSourceConfig.java
@@ -49,7 +49,7 @@ public class RoutingDataSourceConfig {
                     acquireIncrement,
                     "jdbc:mysql://" + config.getDbUrl() + "/" + config.getDbName() + "?zeroDateTimeBehavior=convertToNull&useSSL=false",
                     config.getDbUserName(), config.getDbPassword());
-            targetDataSources.put(config.getId(), comboPooledDataSource);
+            targetDataSources.put((Integer) config.getId().intValue(), comboPooledDataSource);
             routingDataSource.setDefaultTargetDataSource(comboPooledDataSource);
         }
         routingDataSource.setTargetDataSources(targetDataSources);
